Move the external sat and abc prover into the ABC prover module and clean up the interfaces a bit.
